Rename colorbrewer to match fn naming
